requirements.txt
python-dotenv
mcp
git+https://github.com/weatherapicom/python.git#egg=weatherapipython
weatherapi
gradio
langgraph
langchain
langchain-mcp-adapters
langchain-ollama